requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
